New VerboseResult type
The new VerboseResult type has the same structure as the Result type, but some
information about the documents added, e.g. the title and the URI. The new a
annotateResult function can be used to transform a Result into a VerboseResult by
looking up all the document information in the document table of a provided index.

darcs-hash:20080109094158-b81cb-6fddfe2cd09d97f98644487639fc8602274c0333.gz