Remove compression of keys
As the the compression for short lists is quite inefficcient and the keys in the trie tend
to get shorter as more elements are added, the computational overhead is quite
large. Therfore the compression was removed. Additionally, the union functions where
optimized to always insert the small set into the big set.

darcs-hash:20080214124404-67f31-46c3781acb9a7b3fdd8a4b637e6a12f914756d26.gz