New heuristic for measuring word quality
A new heuristic is used for limiting the completions, which should be a little bit faster
and more "correct". Unfortunately the heuristic requires the total amount of documents,
which is not available from the index and has therefore be figured out using the doc
table (which requires transmission of this value to query servers, too).

darcs-hash:20080314131142-67f31-9383c0683b9339b0b5c3be2c3d794ddc35291b49.gz