Index building now based on HolDocuments
Until now, a list of Documents had to be passed as arguments. This has been changed so that HolDocuments are accepted as parameters, transformed into a list and passed to an inner function that is exactly the old one.

darcs-hash:20080610104604-ff21e-0ac326767ab26c67b92e473d7cdd8c25e4e39e97.gz